refactor(demo): refactor demo classes
