+++ enhanced-grid-flow-demo/src/main/java/com/vaadin/componentfactory/enhancedgrid/MainLayout.java
@@ -9,11 +9,13 @@ public class MainLayout extends AppLayout {
 
     public MainLayout() {
         final DrawerToggle drawerToggle = new DrawerToggle();
-        final RouterLink simple = new RouterLink("Multi Selection Grid", SimpleView.class);
-        final RouterLink singleSimple = new RouterLink("SingleSelection Grid", SimpleSingleSelectView.class);
+        
+        final RouterLink simpleMultiGrid = new RouterLink("Multiple Selection Grid", SimpleMultiSelectView.class);
+        final RouterLink simpleSingleGrid = new RouterLink("Single Selection Grid", SimpleSingleSelectView.class);           
+        final RouterLink multiTree = new RouterLink("Multiple Selection Tree Grid", MultiTreeGridView.class);
         final RouterLink singleTree = new RouterLink("Single Selection Tree Grid", SingleTreeGridView.class);
-        final RouterLink multiTree = new RouterLink("Multi Selection Tree Grid", MultiTreeGridView.class);
-        final VerticalLayout menuLayout = new VerticalLayout(simple, singleSimple, singleTree, multiTree);
+        
+        final VerticalLayout menuLayout = new VerticalLayout(simpleMultiGrid, simpleSingleGrid, multiTree, singleTree);
         addToDrawer(menuLayout);
         addToNavbar(drawerToggle);
     }
