+++ enhanced-grid-flow-demo/src/main/java/com/vaadin/componentfactory/enhancedgrid/MultiTreeGridView.java
@@ -1,13 +1,20 @@
 package com.vaadin.componentfactory.enhancedgrid;
 
 import com.vaadin.componentfactory.enhancedgrid.bean.Department;
-import com.vaadin.componentfactory.enhancedgrid.bean.DepartmentData;
+import com.vaadin.componentfactory.enhancedgrid.data.DepartmentData;
 import com.vaadin.componentfactory.enhancedtreegrid.EnhancedTreeGrid;
 import com.vaadin.flow.component.grid.Grid.Column;
 import com.vaadin.flow.component.grid.Grid.SelectionMode;
+import com.vaadin.flow.component.grid.editor.Editor;
 import com.vaadin.flow.component.html.Div;
+import com.vaadin.flow.component.textfield.TextField;
+import com.vaadin.flow.data.binder.Binder;
 import com.vaadin.flow.router.Route;
 
+/**
+ * Basic multiple selection treegrid example
+ * 
+ */
 @Route(value = "multi-tree", layout = MainLayout.class)
 public class MultiTreeGridView extends Div {
 
@@ -16,8 +23,11 @@ public class MultiTreeGridView extends Div {
 		 
 		DepartmentData departmentData = new DepartmentData();
     	EnhancedTreeGrid<Department> grid = new EnhancedTreeGrid<>();
+    	
+    	// set selection predicate to indicate which items can be selected
     	grid.setSelectionPredicate(d -> d.getName().startsWith("T"));
     	
+    	// set selection mode
     	grid.setSelectionMode(SelectionMode.MULTI);
     	
     	grid.asMultiSelect().addValueChangeListener(event -> {
@@ -26,12 +36,34 @@ public class MultiTreeGridView extends Div {
             messageDiv.setText(message);
         });
     	
+    	// set items
 		grid.setItems(departmentData.getRootDepartments(),
 		        departmentData::getChildDepartments);
-		grid.addHierarchyColumn(Department::getName)
+		
+		// add columns
+		Column<Department> nameColumn = grid.addHierarchyColumn(Department::getName)
 		        .setHeader("Department Name");
 		Column<Department> managerColumn = grid.addColumn(Department::getManager).setHeader("Manager");				
 		managerColumn.setSortable(true);		
+		
+		// set editable predicate to indicate which items can be edited
+        grid.setEditablePredicate(d -> d.getName().startsWith("T"));        
+        
+        Binder<Department> binder = new Binder<>(Department.class);
+		Editor<Department> editor = grid.getEditor();
+		editor.setBinder(binder);
+		editor.setBuffered(true);
+	     
+		// define editor components for columns
+        TextField nameField = new TextField();
+        binder.bind(nameField, Department::getName, Department::setName);
+        nameColumn.setEditorComponent(nameField);
+               
+        // call edit
+        grid.addItemDoubleClickListener(event -> {
+            grid.editItem(event.getItem());
+            nameField.focus();
+        });                
 		
 		add(grid, messageDiv);
 	 }
