--- enhanced-grid-flow-demo/src/main/java/com/vaadin/componentfactory/enhancedgrid/SimpleView.java
@@ -1,47 +1,0 @@
-package com.vaadin.componentfactory.enhancedgrid;
-
-import com.vaadin.componentfactory.enhancedgrid.bean.Person;
-import com.vaadin.componentfactory.enhancedgrid.service.PersonService;
-import com.vaadin.flow.component.grid.Grid;
-import com.vaadin.flow.component.html.Div;
-import com.vaadin.flow.router.Route;
-
-import java.util.List;
-
-/**
- * Basic example with setItems
- */
-@Route(value = "", layout = MainLayout.class)
-public class SimpleView extends Div {
-
-
-    public SimpleView() {
-        Div messageDiv = new Div();
-
-        List<Person> personList = getItems();
-        EnhancedGrid<Person> grid = new EnhancedGrid<>();
-        grid.setSelectionPredicate(p -> p.getAge() > 18);
-        grid.setItems(personList);
-
-        grid.addColumn(Person::getFirstName).setHeader("First Name");
-        grid.addColumn(Person::getAge).setHeader("Age");
-
-        grid.setSelectionMode(Grid.SelectionMode.MULTI);
-
-        grid.asMultiSelect().addValueChangeListener(event -> {
-            String message = String.format("Selection changed from %s to %s",
-                event.getOldValue(), event.getValue());
-            messageDiv.setText(message);
-        });
-
-        // You can pre-select items
-        //grid.select(personList.get(1));
-        grid.asMultiSelect().select(personList.get(0), personList.get(1));
-        add(grid, messageDiv);
-    }
-
-    private List<Person> getItems() {
-        PersonService personService = new PersonService();
-        return personService.fetchAll();
-    }
-}
