+++ enhanced-grid-flow-demo/src/main/java/com/vaadin/componentfactory/enhancedgrid/SingleTreeGridView.java
@@ -1,13 +1,21 @@
 package com.vaadin.componentfactory.enhancedgrid;
 
 import com.vaadin.componentfactory.enhancedgrid.bean.Department;
-import com.vaadin.componentfactory.enhancedgrid.bean.DepartmentData;
+import com.vaadin.componentfactory.enhancedgrid.bean.Person;
+import com.vaadin.componentfactory.enhancedgrid.data.DepartmentData;
 import com.vaadin.componentfactory.enhancedtreegrid.EnhancedTreeGrid;
 import com.vaadin.flow.component.grid.Grid.Column;
 import com.vaadin.flow.component.grid.Grid.SelectionMode;
+import com.vaadin.flow.component.grid.editor.Editor;
 import com.vaadin.flow.component.html.Div;
+import com.vaadin.flow.component.textfield.TextField;
+import com.vaadin.flow.data.binder.Binder;
 import com.vaadin.flow.router.Route;
 
+/**
+ * Basic single selection treegrid example
+ * 
+ */
 @Route(value = "single-tree", layout = MainLayout.class)
 public class SingleTreeGridView extends Div {
 
@@ -16,8 +24,11 @@ public class SingleTreeGridView extends Div {
 		 
 		DepartmentData departmentData = new DepartmentData();
     	EnhancedTreeGrid<Department> grid = new EnhancedTreeGrid<>();
+    	
+    	// set selection predicate to indicate which items can be selected
     	grid.setSelectionPredicate(d -> d.getName().startsWith("T"));
     	
+    	// set selection mode
     	grid.setSelectionMode(SelectionMode.SINGLE);
     	
     	grid.asSingleSelect().addValueChangeListener(event -> {
@@ -26,12 +37,34 @@ public class SingleTreeGridView extends Div {
             messageDiv.setText(message);
         });
     	
+    	// set items
 		grid.setItems(departmentData.getRootDepartments(),
 		        departmentData::getChildDepartments);
-		grid.addHierarchyColumn(Department::getName)
+		
+		// add columns
+		Column<Department> nameColumn = grid.addHierarchyColumn(Department::getName)
 		        .setHeader("Department Name");
 		Column<Department> managerColumn = grid.addColumn(Department::getManager).setHeader("Manager");				
 		managerColumn.setSortable(true);		
+		
+		// set editable predicate to indicate which items can be edited
+        grid.setEditablePredicate(d -> d.getName().startsWith("T"));        
+        
+        Binder<Department> binder = new Binder<>(Department.class);
+		Editor<Department> editor = grid.getEditor();
+		editor.setBinder(binder);
+		editor.setBuffered(true);
+	     
+		// define editor components for columns
+        TextField nameField = new TextField();
+        binder.bind(nameField, Department::getName, Department::setName);
+        nameColumn.setEditorComponent(nameField);
+               
+        // call edit
+        grid.addItemDoubleClickListener(event -> {
+            grid.editItem(event.getItem());
+            nameField.focus();
+        });                
 		
 		add(grid, messageDiv);
 	 }
